% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_tips_to_form}
\alias{convert_tips_to_form}
\title{Title}
\usage{
convert_tips_to_form(games_tbl, form, comp)
}
\arguments{
\item{games_tbl}{}

\item{form}{}

\item{comp}{}
}
\value{

}
\description{
Title
}
