% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_current_games}
\alias{get_current_games}
\title{Title}
\usage{
get_current_games(user, pass, comp)
}
\arguments{
\item{user}{}

\item{pass}{}

\item{comp}{}

\item{round}{}
}
\value{

}
\description{
Title
}
