% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_games}
\alias{get_games}
\title{Get Games}
\usage{
get_games(user, pass, comp, round = NULL)
}
\arguments{
\item{user}{monash username, in text}

\item{pass}{monash password, in text}

\item{comp}{comp type, should be one of "normal", "gauss" or "info"}

\item{round}{(optional),  round number to return. If not provided, will try find the current round.}
}
\value{
A tibble, containing the matches from the currently available tipping round
}
\description{
get_current_games provides an API to return the current rounds matches in a data frame.
}
\examples{
\dontrun{
get_games(user, pass, comp)
}
}
