% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_round_games}
\alias{get_round_games}
\title{Title}
\usage{
get_round_games(user, pass, round, comp)
}
\arguments{
\item{user}{}

\item{pass}{}

\item{round}{}

\item{comp}{}
}
\value{

}
\description{
Title
}
