% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{make_request}
\alias{make_request}
\title{Title}
\usage{
make_request(user = NULL, pass = NULL, round = NULL, comp = NULL)
}
\arguments{
\item{user}{}

\item{pass}{}

\item{round}{}

\item{comp}{}
}
\value{

}
\description{
Title
}
