% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monashtipr-package.R
\docType{package}
\name{monashtipr-package}
\alias{monashtipr}
\alias{monashtipr-package}
\title{monashtipr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\keyword{internal}
