% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{submit_tips}
\alias{submit_tips}
\title{Title}
\usage{
submit_tips(games, user, pass, round, comp)
}
\arguments{
\item{games}{}

\item{user}{}

\item{pass}{}

\item{round}{}

\item{comp}{}
}
\value{

}
\description{
Title
}
