% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monashtipr-package.R
\docType{package}
\name{monashtipr-package}
\alias{monashtipr}
\alias{monashtipr-package}
\title{monashtipr: An R API Wrapper for the Monash University Probabilistic Footy Tipping Competition}
\description{
An API wrapper for the 'Monash University Probabilistic Footy 
    Tipping Competition' <https://probabilistic-footy.monash.edu/~footy/index.shtml>. 
    Allows users to submit tips directly to the competition from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jimmyday12.github.io/monash_tipr/}
  \item \url{https://github.com/jimmyday12/monash_tipr}
}

}
\author{
\strong{Maintainer}: James Day \email{jamesthomasday@gmail.com} (\href{https://orcid.org/0000-0003-3133-943X}{ORCID}) [copyright holder]

}
\keyword{internal}
