% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{submit_tips}
\alias{submit_tips}
\title{Submit Tips}
\usage{
submit_tips(games, user, pass, comp, round = NULL)
}
\arguments{
\item{games}{a table of games, ideally returned from \code{get_games}}

\item{user}{monash username, in text}

\item{pass}{monash password, in text}

\item{comp}{comp type, should be one of "normal", "gauss" or "info"}

\item{round}{(optional),  round number to return. If not provided, will try find the current round.}
}
\value{
A tibble, containing the games and tips submitted via the API with a status column
}
\description{
Submit Tips
}
\examples{
\dontrun{
submit_tips(games, user, pass, comp)
}
}
